requirements.txt
discord
pandas
pynacl
python-dotenv
pymongo